Refactor search, escape query strings for Tantivy
Revert "Add betafeatures"

This reverts commit 87145e9c30e63a516739c342b42111d045d516de.
